Add support for audio sampling
Extends the ADC12B driver to support using TC2 to sample continuously at
a fixed rate. Samples are batch-requested by application code, and an
overrun flag is provided.

A demo app is used to check basic functionality. The over-flow can be
checked by modifying the call to Adc12GetSamples() to only request 1
sample.