Rework USB driver to be simpler
Removes much of the higher-layer handling that was happening
automatically, particularly around descriptors. Instead types with
layouts that match the raw binary (assuming GCC-arm) are provided
along with some convenience functions for easily sending them in
responses.

The benefits are:
- Less RAM usage by the driver (less to track). The equivalent tracking
  in the app can be less generic and will likely be more efficient.
- (Some) descriptors can be declared fully in ROM.
- App has more control over things like exact string IDs and interface
  numbers, which is important for some device classes like audio.

The downside is there is less verification that the app is putting the
right things in descriptors. Some typing has been used to try to make
the allowed values somewhat obvious.