Reduce debug driver cpu usage
Once the debug driver had echoed 16 characters it was checking each of
it's 16 message tokens every tick. This may not seem like a lot but each
check involves looping through 64 messages. Even assuming a conservative
10 instructions per loop that gives 10 * 16 * 64 == 20K instructions,
which is ~1/3 of the available cycles per ms.

Using the profiling prints shows that this optimization reduces driver
cpu usage from ~33% to ~3% in the case where at least 16 characters have
been echoed back to the terminal.