Implement USB DMA transfers, setup audio pipeline
Add API for using DMA with USB. Use this for the audio endpoints.
Restructure the audio code into an actual pipeline. The current setup
can get hung up if the audio input is closed by the host while a DMA
read is pending, as the DMA transfer is stuck in this case. A follow up
change will be done to address this.